SAGLIK_PROJESI/requirements.txt
# Mevcut kütüphaneler
folium
geopy
pyttsx3
SpeechRecognition
opencv-python
googletrans==4.0.0-rc1
gTTS
pytesseract
streamlit==1.10.0
pandas==1.3.5
numpy==1.21.6
plotly==5.3.1
scikit-learn==1.0.2
joblib==1.1.0
datetime==4.3
# Güvenlik kütüphaneleri
PyJWT
bcrypt
cryptography
